a very important bug in serialization has been found and fixed, see description
When values are written as array in a file, since value just has one token, such as xyz={ value1 value2 }, in previous parser, value1 and value2 will be parsed to element only has name, and its name is the value (say value1, value2), which will cause a problem.

That is, since when append a element to ElmentScope, the latter will classify the former by name, and take elements has same name together into one list.

Because values in array has order in most time, classify them by name will change the order when an array contains some same values. In such case, now the  parser will treat a value as an element has name of "", and uses tag to present the value, so that when appending to ElementScope, all values will treat as nameless element and put in one list together in their raw order.